new thread to download ftp
Now monitor mail to get ftp_info is a thread
and download the ftp_info is another thread

the first thread put the ftp_info into a list called self.ftp_queue, and the download thread will get the item from this list.
Also, the GUI has now a label to display the queued ftp_info number with the filtering feature that dosen't allow the repeated ftp_info to be queued.